+++ lib/collection/src/recommendations.rs
@@ -1,21 +1,17 @@
 use std::future::Future;
-use std::iter::Peekable;
 use std::sync::Arc;
 use std::time::Duration;
 
 use api::rest::RecommendStrategy;
 use common::counter::hardware_accumulator::HwMeasurementAcc;
 use itertools::Itertools;
-use segment::data_types::vectors::{
-    DenseVector, NamedQuery, TypedMultiDenseVector, VectorElementType, VectorInternal, VectorRef,
-};
+use segment::data_types::vectors::{NamedQuery, VectorInternal};
 use segment::types::{
     Condition, ExtendedPointId, Filter, HasIdCondition, PointIdType, ScoredPoint,
 };
-use segment::vector_storage::query::RecoQuery;
+use segment::vector_storage::query::{RecoQuery, avg_vector_for_recommendation};
 use shard::query::query_enum::QueryEnum;
 use shard::search::CoreSearchRequestBatch;
-use sparse::common::sparse_vector::SparseVector;
 
 use crate::collection::Collection;
 use crate::common::batching::{batch_requests, empty_batch_results};
@@ -30,120 +26,6 @@ use crate::operations::shard_selector_internal::ShardSelectorInternal;
 use crate::operations::types::{
     CollectionError, CollectionResult, CoreSearchRequest, RecommendRequestInternal, UsingVector,
 };
-
-fn avg_vectors<'a>(
-    vectors: impl IntoIterator<Item = VectorRef<'a>>,
-) -> CollectionResult<VectorInternal> {
-    let mut avg_dense = DenseVector::default();
-    let mut avg_sparse = SparseVector::default();
-    let mut avg_multi: Option<TypedMultiDenseVector<VectorElementType>> = None;
-    let mut dense_count = 0;
-    let mut sparse_count = 0;
-    let mut multi_count = 0;
-    for vector in vectors {
-        match vector {
-            VectorRef::Dense(vector) => {
-                dense_count += 1;
-                for i in 0..vector.len() {
-                    if i >= avg_dense.len() {
-                        avg_dense.push(vector[i])
-                    } else {
-                        avg_dense[i] += vector[i];
-                    }
-                }
-            }
-            VectorRef::Sparse(vector) => {
-                sparse_count += 1;
-                avg_sparse = vector.combine_aggregate(&avg_sparse, |v1, v2| v1 + v2);
-            }
-            VectorRef::MultiDense(vector) => {
-                multi_count += 1;
-                avg_multi = Some(avg_multi.map_or_else(
-                    || vector.to_owned(),
-                    |mut avg_multi| {
-                        avg_multi
-                            .flattened_vectors
-                            .extend_from_slice(vector.flattened_vectors);
-                        avg_multi
-                    },
-                ));
-            }
-        }
-    }
-
-    match (dense_count, sparse_count, multi_count) {
-        // TODO(sparse): what if vectors iterator is empty? We added CollectionError::BadRequest,
-        // but it's not clear if it's the best solution.
-        // Currently it's hard to return an zeroed vector, because we don't know its type: dense or sparse.
-        (0, 0, 0) => Err(CollectionError::bad_input(
-            "Positive vectors should not be empty with `average` strategy".to_owned(),
-        )),
-        (_, 0, 0) => {
-            for item in &mut avg_dense {
-                *item /= dense_count as VectorElementType;
-            }
-            Ok(VectorInternal::from(avg_dense))
-        }
-        (0, _, 0) => {
-            for item in &mut avg_sparse.values {
-                *item /= sparse_count as VectorElementType;
-            }
-            Ok(VectorInternal::from(avg_sparse))
-        }
-        (0, 0, _) => match avg_multi {
-            Some(avg_multi) => Ok(VectorInternal::from(avg_multi)),
-            None => Err(CollectionError::bad_input(
-                "Positive vectors should not be empty with `average` strategy".to_owned(),
-            )),
-        },
-        (_, _, _) => Err(CollectionError::bad_input(
-            "Can't average vectors with different types".to_owned(),
-        )),
-    }
-}
-
-fn merge_positive_and_negative_avg(
-    positive: VectorInternal,
-    negative: VectorInternal,
-) -> CollectionResult<VectorInternal> {
-    match (positive, negative) {
-        (VectorInternal::Dense(positive), VectorInternal::Dense(negative)) => {
-            let vector: DenseVector = positive
-                .iter()
-                .zip(negative.iter())
-                .map(|(pos, neg)| pos + pos - neg)
-                .collect();
-            Ok(vector.into())
-        }
-        (VectorInternal::Sparse(positive), VectorInternal::Sparse(negative)) => Ok(positive
-            .combine_aggregate(&negative, |pos, neg| pos + pos - neg)
-            .into()),
-        (VectorInternal::MultiDense(mut positive), VectorInternal::MultiDense(negative)) => {
-            // merge positive and negative vectors as concatenated vectors with negative vectors negated
-            positive.flattened_vectors.extend(negative.flattened_vectors.into_iter().map(|x| -x));
-            Ok(VectorInternal::MultiDense(positive))
-        },
-        _ => Err(CollectionError::bad_input(
-            "Positive and negative vectors should be of the same type, either all dense or all sparse or all multi".to_owned(),
-        )),
-    }
-}
-
-pub fn avg_vector_for_recommendation<'a>(
-    positive: impl IntoIterator<Item = VectorRef<'a>>,
-    mut negative: Peekable<impl Iterator<Item = VectorRef<'a>>>,
-) -> CollectionResult<VectorInternal> {
-    let avg_positive = avg_vectors(positive)?;
-
-    let search_vector = if negative.peek().is_none() {
-        avg_positive
-    } else {
-        let avg_negative = avg_vectors(negative)?;
-        merge_positive_and_negative_avg(avg_positive, avg_negative)?
-    };
-
-    Ok(search_vector)
-}
 
 #[allow(clippy::too_many_arguments)]
 pub async fn recommend_by<F, Fut>(
@@ -469,49 +351,5 @@ fn recommend_by_custom_score(
         with_payload,
         with_vector,
         score_threshold,
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use segment::data_types::vectors::{VectorInternal, VectorRef};
-    use sparse::common::sparse_vector::SparseVector;
-
-    use super::avg_vectors;
-
-    #[test]
-    fn test_avg_vectors() {
-        let vectors: Vec<VectorInternal> = vec![
-            vec![1.0, 2.0, 3.0].into(),
-            vec![1.0, 2.0, 3.0].into(),
-            vec![1.0, 2.0, 3.0].into(),
-        ];
-        assert_eq!(
-            avg_vectors(vectors.iter().map(VectorRef::from)).unwrap(),
-            vec![1.0, 2.0, 3.0].into(),
-        );
-
-        let vectors: Vec<VectorInternal> = vec![
-            SparseVector::new(vec![0, 1, 2], vec![0.0, 0.1, 0.2])
-                .unwrap()
-                .into(),
-            SparseVector::new(vec![0, 1, 2], vec![0.0, 1.0, 2.0])
-                .unwrap()
-                .into(),
-        ];
-        assert_eq!(
-            avg_vectors(vectors.iter().map(VectorRef::from)).unwrap(),
-            SparseVector::new(vec![0, 1, 2], vec![0.0, 0.55, 1.1])
-                .unwrap()
-                .into(),
-        );
-
-        let vectors: Vec<VectorInternal> = vec![
-            vec![1.0, 2.0, 3.0].into(),
-            SparseVector::new(vec![0, 1, 2], vec![0.0, 0.1, 0.2])
-                .unwrap()
-                .into(),
-        ];
-        assert!(avg_vectors(vectors.iter().map(VectorRef::from)).is_err());
     }
 }
